First push containing the C# code. So far its only half of the add folders program.
Its amazing at the difference in changing languages does. The code is almost half the size compared to Java, granted I probably didn't optimize the code as well as I could of.

Being that Java is the first language I learned, and I'm still learning it, it's a lot faster to write than C#. I haven't done anything in C# for 2 years since I touched Silverlight for an internship, where I didn't do anything with Directory or file manipulations so i have to do a lot of Googling and probably finding the most streamlined way to do things.