More work on the main program
Did a little bit of work while stuck sitting in the computer lab where I work part time. Kind of half-assed at going it. so theres not that much in there.

Ok its a bit of a lie, I got most of the Main file completed and about a 3rd of the class file.