Got most of the rest of the code converted. Major parts still missing.
The parts i have left are a little more than just looking at the Java and converting it over. The regex part needs to be tested before going any farther. Next then is the massive googling for the file manipulations. I'm sure they are easy in comparison to Java, probably as simple as File.Copy and File.Rename but until I'm sure of that I get no joy in swinging a club blindly.