Version 2.0 completed!!! (almost)
Finally finished the rewrite in C#. There is no change in how the C# works compared to the Java. The only major thing is single files hanging out in the Starting Directory for the sorting are being passed over currently.

I have 3 ways I've thought about fixing it. First is to have it rescan the Starting Directory after clearing out folders and deleting them, problem folders not likely to be deleted if an issues arises in the copy or renaming. The second is to get it to work by creating a condition where it will search there without going through the motions if files exist(should be that way already), problem is thats more like patching the issues with a big bandage that can fail in the event of a problem. Third is to get it to work the proper way, make sure it does not do nameStripping if originalFiles is empty, and to clear originalFiles after sorting.

Its also to note theres not much difference in lines of code either with the finished copy. The C# could probably be simplified more, but performance its probably as close to perfect as it'll get, unless overuse of lambda's/LINQ are faster.