Fixed and finished the recursion, V2.0 finally considered done.
I consider Version 2.0 to be done, even though it still has alot of work to do to fix minor annoyances.

Annoyances:
.dat files still exist and muck up the program.
does not differentiate between extensions
doesnt actually create a log of failed files when trying to move.
possibility of it being sped up (i.e. getting a full path to all files into a giant list and sorting through that way, not sure how much of a speed improvement that would leave)